+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Changes
 * [UI]: Removed phylocanvas context menu from analysis > details page.
 * [UI]: Added buttons on the project settings page, and metadata template page to download an excel template.
 * [Developer]: Reorganized permissions classes into subpackages.
+* [UI]: Added maximum coverage quality control option.
 
 0.16.0 to 0.17.0
 ----------------